feat(build): migrate from SCSS to Tailwind CSS 4
- Add `@tailwindcss/vite` plugin and configuration.
- Replace `scripts/gen-colors.ts` with `src/bin/gen-theme.ts` to generate CSS custom properties instead of SCSS maps.
- Initialize `src/styles/index.css` and `src/styles/theme.css` for centralized theme management.
- Update `package.json` to reflect the new theme generation script and necessary dependency upgrades for Tailwind 4 support.
- Add project-specific Zed settings for improved CSS language server integration.

Why: Migrating to Tailwind CSS 4 provides better utility-first styling and modernized theme integration.
