+++ .zed/settings.json
@@ -1,0 +1,11 @@
+// Folder-specific settings
+//
+// For a full list of overridable settings, and general information on folder-specific settings,
+// see the documentation: https://zed.dev/docs/configuring-zed#settings-files
+{
+	"languages": {
+		"CSS": {
+			"language_servers": ["tailwindcss-intellisense-css"]
+		}
+	}
+}
