+++ packages/svelte/package.json
@@ -22,7 +22,7 @@
 		"format": "prettier --write .",
 		"lint": "prettier --check . && eslint .",
 		"npm-publish": "bun run build && export $(grep -v '^#' ../../.env | xargs) && npm publish --access=public",
-		"gen-colors": "bun run ./scripts/gen-colors.ts"
+		"gen-theme": "bun run ./src/bin/gen-theme.ts"
 	},
 	"files": [
 		"dist",
@@ -53,30 +53,31 @@
 	"devDependencies": {
 		"@eslint/compat": "^1.4.1",
 		"@eslint/js": "^9.39.4",
-		"@iconify-json/hugeicons": "^1.2.24",
+		"@iconify-json/hugeicons": "^1.2.25",
 		"@sveltejs/adapter-auto": "^4.0.0",
-		"@sveltejs/kit": "^2.55.0",
+		"@sveltejs/kit": "^2.57.1",
 		"@sveltejs/package": "^2.5.7",
 		"@sveltejs/vite-plugin-svelte": "^5.1.1",
-		"@types/bun": "^1.3.10",
+		"@tailwindcss/vite": "^4.2.2",
 		"autoprefixer": "^10.4.27",
-		"bits-ui": "^2.16.3",
+		"bits-ui": "^2.17.3",
 		"csstype": "^3.2.3",
 		"eslint": "^9.39.4",
 		"eslint-config-prettier": "^10.1.8",
-		"eslint-plugin-svelte": "^3.15.2",
+		"eslint-plugin-svelte": "^3.17.0",
 		"globals": "^16.5.0",
 		"material-dynamic-colors": "^1.1.4",
-		"prettier": "^3.8.1",
+		"prettier": "^3.8.2",
 		"prettier-plugin-svelte": "^3.5.1",
 		"publint": "^0.3.18",
-		"sass": "^1.98.0",
-		"svelte": "^5.53.12",
-		"svelte-check": "^4.4.5",
+		"sass": "^1.99.0",
+		"svelte": "^5.55.3",
+		"svelte-check": "^4.4.6",
+		"tailwindcss": "^4.2.2",
 		"typescript": "^5.9.3",
-		"typescript-eslint": "^8.57.0",
+		"typescript-eslint": "^8.58.1",
 		"unplugin-icons": "^22.5.0",
-		"vite": "^6.4.1"
+		"vite": "^6.4.2"
 	},
 	"keywords": [
 		"svelte"
