+++ packages/svelte/src/bin/gen-theme.ts
@@ -1,0 +1,58 @@
+#!/usr/bin/env bun
+
+/**
+ * Generate Tailwind 4 @theme CSS from Material Color Utilities theme output.
+ * Usage: bun index.ts <output-path>
+ *
+ * This script generates a CSS file with @variant and @theme directives,
+ * using the output of Material Color Utilities' themeFromSourceColor.
+ */
+
+import materialDynamicColors from 'material-dynamic-colors';
+import { writeFile } from 'node:fs/promises';
+import path from 'node:path';
+
+// ---- CONFIGURABLE SOURCE COLOR / OUTPUT ----
+const argv = process.argv.slice(2);
+const getFlag = (names: string[]) => {
+	const i = argv.findIndex((a) => names.includes(a));
+	if (i >= 0 && argv[i + 1] && !argv[i + 1].startsWith('--')) return argv[i + 1];
+	const kv = argv.find((a) => names.some((n) => a.startsWith(`${n}=`)));
+	return kv ? kv.split('=').slice(1).join('=') : undefined;
+};
+const sourceColor = (getFlag(['--color', '-c']) || process.env.SOURCE_COLOR || '#ff907f').trim();
+const outputPath = path.resolve(
+	process.cwd(),
+	getFlag(['--out', '-o']) || process.env.OUTPUT_PATH || 'src/styles/theme.css'
+);
+
+// ---- GENERATE THEME ----
+const theme = await materialDynamicColors(sourceColor);
+
+// ---- TAILWIND 4 @THEME CSS OUTPUT ----
+function toThemeVariables(colors: Record<string, string>): string {
+	return Object.entries(colors)
+		.map(([k, v]) => `  --color-${k}: ${v};`)
+		.join('\n');
+}
+
+const lightVars = toThemeVariables(theme.light as unknown as Record<string, string>);
+const darkVars = toThemeVariables(theme.dark as unknown as Record<string, string>);
+
+const css = `/* Generated by index.ts */
+/* Source color: ${sourceColor} */
+
+@custom-variant dark (&:where([data-theme=dark], [data-theme=dark] *));
+
+@theme {
+${lightVars}
+}
+
+@variant dark {
+${darkVars}
+}
+`;
+
+// ---- WRITE FILE ----
+await writeFile(outputPath, css);
+console.log(`Material theme CSS written to ${outputPath}`);
