+++ packages/svelte/src/styles/index.css
@@ -1,0 +1,2 @@
+@import 'tailwindcss';
+@import './theme.css';
