+++ packages/svelte/src/styles/theme.css
@@ -1,0 +1,82 @@
+/* Generated by index.ts */
+/* Source color: #ff907f */
+
+@custom-variant dark (&:where([data-theme=dark], [data-theme=dark] *));
+
+@theme {
+  --color-primary: #9c4235;
+  --color-onPrimary: #ffffff;
+  --color-primaryContainer: #ffdad4;
+  --color-onPrimaryContainer: #410000;
+  --color-secondary: #775651;
+  --color-onSecondary: #ffffff;
+  --color-secondaryContainer: #ffdad4;
+  --color-onSecondaryContainer: #2c1512;
+  --color-tertiary: #705c2e;
+  --color-onTertiary: #ffffff;
+  --color-tertiaryContainer: #fbdfa6;
+  --color-onTertiaryContainer: #251a00;
+  --color-error: #ba1a1a;
+  --color-onError: #ffffff;
+  --color-errorContainer: #ffdad6;
+  --color-onErrorContainer: #410002;
+  --color-background: #fffbff;
+  --color-onBackground: #201a19;
+  --color-surface: #fff8f6;
+  --color-onSurface: #201a19;
+  --color-surfaceVariant: #f5ddda;
+  --color-onSurfaceVariant: #534341;
+  --color-outline: #857370;
+  --color-outlineVariant: #d8c2be;
+  --color-shadow: #000000;
+  --color-scrim: #000000;
+  --color-inverseSurface: #362f2e;
+  --color-inverseOnSurface: #fbeeec;
+  --color-inversePrimary: #ffb4a8;
+  --color-surfaceDim: #e4d7d5;
+  --color-surfaceBright: #fff8f6;
+  --color-surfaceContainerLowest: #ffffff;
+  --color-surfaceContainerLow: #fef1ee;
+  --color-surfaceContainer: #f8ebe9;
+  --color-surfaceContainerHigh: #f3e5e3;
+  --color-surfaceContainerHighest: #ede0dd;
+}
+
+@variant dark {
+  --color-primary: #ffb4a8;
+  --color-onPrimary: #5f150d;
+  --color-primaryContainer: #7d2b21;
+  --color-onPrimaryContainer: #ffdad4;
+  --color-secondary: #e7bdb6;
+  --color-onSecondary: #442925;
+  --color-secondaryContainer: #5d3f3b;
+  --color-onSecondaryContainer: #ffdad4;
+  --color-tertiary: #dec48c;
+  --color-onTertiary: #3e2e04;
+  --color-tertiaryContainer: #564419;
+  --color-onTertiaryContainer: #fbdfa6;
+  --color-error: #ffb4ab;
+  --color-onError: #690005;
+  --color-errorContainer: #93000a;
+  --color-onErrorContainer: #ffb4ab;
+  --color-background: #201a19;
+  --color-onBackground: #ede0dd;
+  --color-surface: #181211;
+  --color-onSurface: #ede0dd;
+  --color-surfaceVariant: #534341;
+  --color-onSurfaceVariant: #d8c2be;
+  --color-outline: #a08c89;
+  --color-outlineVariant: #534341;
+  --color-shadow: #000000;
+  --color-scrim: #000000;
+  --color-inverseSurface: #ede0dd;
+  --color-inverseOnSurface: #362f2e;
+  --color-inversePrimary: #9c4235;
+  --color-surfaceDim: #181211;
+  --color-surfaceBright: #3f3736;
+  --color-surfaceContainerLowest: #120d0c;
+  --color-surfaceContainerLow: #201a19;
+  --color-surfaceContainer: #251e1d;
+  --color-surfaceContainerHigh: #2f2827;
+  --color-surfaceContainerHighest: #3b3332;
+}
