+++ packages/svelte/vite.config.ts
@@ -2,10 +2,12 @@ import { sveltekit } from '@sveltejs/kit/vite';
 import { defineConfig } from 'vite';
 import autoprefixer from 'autoprefixer';
 import Icons from 'unplugin-icons/vite';
+import tailwindcss from '@tailwindcss/vite';
 
 export default defineConfig({
 	plugins: [
 		sveltekit(),
+		tailwindcss(),
 		Icons({
 			compiler: 'svelte'
 		})
